Restructure based on feedback
